Admin2

+++ app/Http/Controllers/PersonneController.php
@@ -4,20 +4,59 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\User;
+use Illuminate\Support\Facades\Auth;
 
 class PersonneController extends Controller
 {
     public function getAllPersonne(){
-        $users = User::all();
-        return view('allPersonnes')->with('users', $users);
+
+        if (Auth::check()) {
+            if (Auth::user()->admin) {
+
+                $users = User::all();
+                return view('allPersonnes')->with(compact('users'));
+            }
+            else {
+                return redirect()->action('ToolsController@main');   
+            }
+        }
+        else {
+            return redirect()->action('ToolsController@main');   
+        }
     }
 
     public function getPersonne($id){
-        $user = User::where('id', $id)->first();
-        return view('modifPersonne')->with('user', $user);
+
+        if (Auth::check()) {
+            if (Auth::user()->admin) {
+                
+                $user = User::find($id);
+                if ($user == null) {
+                    return redirect()->action('ToolsController@main');  
+                }
+                else {
+                    return view('modifPersonne')->with('user', $user);
+                }
+            }
+            else {
+                return redirect()->action('ToolsController@main');   
+            }
+        }
+        else {
+            return redirect()->action('ToolsController@main');   
+        }
     }
 
     public function getPersonneEx($id, Request $request){
+
+        $request->validate([
+            'nom'=> 'required|alpha',
+            'prenom'=> 'required|alpha',
+            'mail'=> 'required|email:rfc',
+            'location'=> '',
+            'admin'=> ''
+        ]);
+
         if($request->input('location')){
             $loc = 1;
         }
@@ -33,31 +72,39 @@ class PersonneController extends Controller
         }
 
         $personne = [
-            'nom'=>htmlspecialchars($request->input('nom')),
-            'prenom'=>htmlspecialchars($request->input('prenom')),
-            'mail'=>htmlspecialchars($request->input('mail')),
+            'nom'=>$request->input('nom'),
+            'prenom'=>$request->input('prenom'),
+            'mail'=>$request->input('mail'),
             'vehicle'=>$loc,
             'admin'=>$admin
 
         ];
 
-        User::where('id', $id)->first()->update($personne);
+        User::find($id)->update($personne);
         $alert = 'Votre modidication sur un utilisateur à été réalisé avec succès';
         $request->session()->flash('alert_mp', $alert);
         return redirect()->action('ToolsController@main');
     }
 
-
     public function deletePersonne($id, Request $request){
 
-        User::where('id', $id)->delete();
+        if (Auth::check()) {
+            if (Auth::user()->admin) {
+                
+                User::where('id', $id)->delete();
 
-        $alert = 'Votre suppression sur un véhicule à été réalisé avec succès';
-        $request->session()->flash('alert_mp', $alert);
-
-        
-        
-        return redirect()->action('ToolsController@main');
+                $alert = 'Votre suppression sur un véhicule à été réalisé avec succès';
+                $request->session()->flash('alert_mp', $alert);
+                
+                return redirect()->action('ToolsController@main');
+            }
+            else {
+                return redirect()->action('ToolsController@main');   
+            }
+        }
+        else {
+            return redirect()->action('ToolsController@main');   
+        }
 
     }
 }
